Remove ignored files and added some test cases

+++ c_drawing_funcs.c
@@ -39,7 +39,7 @@ uint8_t get_a(uint32_t color){
 }
 
 
-uint64_t set_Nth_bit(uint32_t x, uint32_t n, uint8_t bit) {
+uint32_t set_Nth_bit(uint32_t x, uint32_t n, uint8_t bit) {
   if(bit){
     return x | (1UL<<n);
   }
